core/requirements.main.txt
docker>=4.4.0
aiodocker>=0.19.1
aiohttp[speedups]>=3.7.3
cchardet>=2.1.7
aiodns>=2.0.0
brotlipy>=0.7.0
aiohttp_cors>=0.7.0
grpcio>=1.36.1
asyncpg>=0.22.0
pyyaml>=5.4.1
psutil>=5.8.0
pyjwt>=2.0.1
elasticsearch[async]>=7.0.0,<8.0.0
uvloop>=0.15.2
aioredis>=1.3.1
orjson>=3.5.1
protobuf>=3.15.6
numpy>=1.20.2
colorama>=0.4.4
coloredlogs>=15.0
pycryptodome>=3.10.1
